modifified accessTocken as static member of BaseRestService, added real endpoints call for services, added classes for mocked data from services endpoints
