+++ src/services/base.ts
@@ -2,8 +2,10 @@ import axios from "axios";
 import CONFIGURATION from "@/config";
 
 
+
+
 export default class BaseRestService {
-  accessToken: string | null = null;
+  static accessToken: string | null = null;
 
   restClient = axios.create();
 
@@ -12,16 +14,16 @@ export default class BaseRestService {
     if (CONFIGURATION.auth?.isAuthenticationNeeded) {
       console.debug("application need authentication");
 
-      if (!this.accessToken) {
+      if (!BaseRestService.accessToken) {
         //no authentication token, redirect to login
         console.debug("no token provided... redirecting user to login");
       }
       else {
         //token ok
-        console.debug("token available: ", this.accessToken);
+        console.debug("token available: ", BaseRestService.accessToken);
         //register a transformer for the authentication header
         this.restClient.defaults.transformRequest = (data, headers) => {
-          headers['Authorization'] = `Basic ${this.accessToken}`;
+          headers['Authorization'] = `Basic ${BaseRestService.accessToken}`;
           return data;
         }
         //register response interceptor for unauthorized access
